+++ ush/b2i/bufr2ioda_insitu_surface_ndbc.py
@@ -37,7 +37,7 @@ class NDBCIODAVariables(IODAVariables):
         return q
 
     def set_obs_from_query_result(self, r):
-        self.temp = r.get('temp')
+        self.temp = r.get('temp').astype(np.float32)
         self.temp -= 273.15
 
     def filter(self):
